+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/network/node/NetworkNodeCrafter.java
@@ -115,7 +115,7 @@ public class NetworkNodeCrafter extends NetworkNode implements ICraftingPatternC
         if (!state) {
             network.getCraftingManager().getTasks().stream()
                 .filter(task -> task.getPattern().getContainer().getPosition().equals(pos))
-                .forEach(task -> network.getCraftingManager().cancel(task));
+                .forEach(task -> network.getCraftingManager().cancel(task.getId()));
         }
 
         network.getCraftingManager().rebuild();