+++ src/main/java/com/raoulvdberge/refinedstorage/network/MessageCraftingMonitorCancel.java
@@ -3,32 +3,41 @@ package com.raoulvdberge.refinedstorage.network;
 import com.raoulvdberge.refinedstorage.container.ContainerCraftingMonitor;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.entity.player.EntityPlayerMP;
+import net.minecraftforge.fml.common.network.ByteBufUtils;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 
+import java.util.UUID;
+
 public class MessageCraftingMonitorCancel extends MessageHandlerPlayerToServer<MessageCraftingMonitorCancel> implements IMessage {
-    private int id;
+    private UUID taskId;
 
     public MessageCraftingMonitorCancel() {
     }
 
-    public MessageCraftingMonitorCancel(int id) {
-        this.id = id;
+    public MessageCraftingMonitorCancel(UUID taskId) {
+        this.taskId = taskId;
     }
 
     @Override
     public void fromBytes(ByteBuf buf) {
-        id = buf.readInt();
+        if (buf.readBoolean()) {
+            taskId = UUID.fromString(ByteBufUtils.readUTF8String(buf));
+        }
     }
 
     @Override
     public void toBytes(ByteBuf buf) {
-        buf.writeInt(id);
+        buf.writeBoolean(taskId != null);
+
+        if (taskId != null) {
+            ByteBufUtils.writeUTF8String(buf, taskId.toString());
+        }
     }
 
     @Override
     public void handle(MessageCraftingMonitorCancel message, EntityPlayerMP player) {
         if (player.openContainer instanceof ContainerCraftingMonitor) {
-            ((ContainerCraftingMonitor) player.openContainer).getCraftingMonitor().onCancelled(player, message.id);
+            ((ContainerCraftingMonitor) player.openContainer).getCraftingMonitor().onCancelled(player, message.taskId);
         }
     }
 }