+++ src/main/java/com/raoulvdberge/refinedstorage/tile/craftingmonitor/ICraftingMonitor.java
@@ -1,5 +1,6 @@
 package com.raoulvdberge.refinedstorage.tile.craftingmonitor;
 
+import com.google.common.base.Optional;
 import com.raoulvdberge.refinedstorage.api.autocrafting.ICraftingManager;
 import com.raoulvdberge.refinedstorage.api.autocrafting.task.ICraftingTask;
 import com.raoulvdberge.refinedstorage.tile.data.TileDataParameter;
@@ -7,16 +8,19 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
 
 import javax.annotation.Nullable;
-import java.util.List;
+import java.util.Collection;
+import java.util.UUID;
 
 public interface ICraftingMonitor {
+    int TABS_PER_PAGE = 6;
+
     String getGuiTitle();
 
-    void onCancelled(EntityPlayerMP player, int id);
+    void onCancelled(EntityPlayerMP player, @Nullable UUID id);
 
     TileDataParameter<Integer, ?> getRedstoneModeParameter();
 
-    List<ICraftingTask> getTasks();
+    Collection<ICraftingTask> getTasks();
 
     @Nullable
     ICraftingManager getCraftingManager();
@@ -28,4 +32,12 @@ public interface ICraftingMonitor {
     boolean isActive();
 
     void onClosed(EntityPlayer player);
+
+    Optional<UUID> getTabSelected();
+
+    int getTabPage();
+
+    void onTabSelectionChanged(Optional<UUID> taskId);
+
+    void onTabPageChanged(int page);
 }