+++ src/main/java/com/raoulvdberge/refinedstorage/tile/craftingmonitor/TileCraftingMonitor.java
@@ -1,5 +1,6 @@
 package com.raoulvdberge.refinedstorage.tile.craftingmonitor;
 
+import com.google.common.base.Optional;
 import com.raoulvdberge.refinedstorage.api.network.grid.IGrid;
 import com.raoulvdberge.refinedstorage.apiimpl.network.node.NetworkNodeCraftingMonitor;
 import com.raoulvdberge.refinedstorage.gui.GuiBase;
@@ -11,6 +12,7 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
 import javax.annotation.Nonnull;
+import java.util.UUID;
 
 public class TileCraftingMonitor extends TileNode<NetworkNodeCraftingMonitor> {
     public static final TileDataParameter<Integer, TileCraftingMonitor> SIZE = new TileDataParameter<>(DataSerializers.VARINT, IGrid.SIZE_STRETCH, t -> t.getNode().getSize(), (t, v) -> {
@@ -19,9 +21,26 @@ public class TileCraftingMonitor extends TileNode<NetworkNodeCraftingMonitor> {
             t.getNode().markDirty();
         }
     }, (initial, p) -> GuiBase.executeLater(GuiCraftingMonitor.class, GuiBase::initGui));
+    public static final TileDataParameter<Optional<UUID>, TileCraftingMonitor> TAB_SELECTED = new TileDataParameter<>(DataSerializers.OPTIONAL_UNIQUE_ID, Optional.absent(), t -> t.getNode().getTabSelected(), (t, v) -> {
+        if (v.isPresent() && t.getNode().getTabSelected().isPresent() && v.get().equals(t.getNode().getTabSelected().get())) {
+            t.getNode().setTabSelected(Optional.absent());
+        } else {
+            t.getNode().setTabSelected(v);
+        }
+
+        t.getNode().markDirty();
+    });
+    public static final TileDataParameter<Integer, TileCraftingMonitor> TAB_PAGE = new TileDataParameter<>(DataSerializers.VARINT, 0, t -> t.getNode().getTabPage(), (t, v) -> {
+        if (v >= 0) {
+            t.getNode().setTabPage(v);
+            t.getNode().markDirty();
+        }
+    });
 
     public TileCraftingMonitor() {
         dataManager.addWatchedParameter(SIZE);
+        dataManager.addWatchedParameter(TAB_SELECTED);
+        dataManager.addWatchedParameter(TAB_PAGE);
     }
 
     @Override