+++ src/main/java/com/raoulvdberge/refinedstorage/tile/craftingmonitor/WirelessCraftingMonitor.java
@@ -1,5 +1,6 @@
 package com.raoulvdberge.refinedstorage.tile.craftingmonitor;
 
+import com.google.common.base.Optional;
 import com.raoulvdberge.refinedstorage.RS;
 import com.raoulvdberge.refinedstorage.api.autocrafting.ICraftingManager;
 import com.raoulvdberge.refinedstorage.api.autocrafting.task.ICraftingTask;
@@ -7,7 +8,7 @@ import com.raoulvdberge.refinedstorage.api.network.INetwork;
 import com.raoulvdberge.refinedstorage.gui.GuiBase;
 import com.raoulvdberge.refinedstorage.gui.GuiCraftingMonitor;
 import com.raoulvdberge.refinedstorage.item.ItemWirelessCraftingMonitor;
-import com.raoulvdberge.refinedstorage.network.MessageWirelessCraftingMonitorSize;
+import com.raoulvdberge.refinedstorage.network.MessageWirelessCraftingMonitorSettings;
 import com.raoulvdberge.refinedstorage.tile.data.TileDataParameter;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
@@ -18,8 +19,9 @@ import net.minecraft.world.World;
 import net.minecraftforge.common.DimensionManager;
 
 import javax.annotation.Nullable;
+import java.util.Collection;
 import java.util.Collections;
-import java.util.List;
+import java.util.UUID;
 
 public class WirelessCraftingMonitor implements ICraftingMonitor {
     private ItemStack stack;
@@ -27,12 +29,16 @@ public class WirelessCraftingMonitor implements ICraftingMonitor {
     private int networkDimension;
     private BlockPos network;
     private int size;
+    private int tabPage;
+    private Optional<UUID> tabSelected;
 
     public WirelessCraftingMonitor(int networkDimension, ItemStack stack) {
         this.stack = stack;
         this.networkDimension = networkDimension;
         this.network = new BlockPos(ItemWirelessCraftingMonitor.getX(stack), ItemWirelessCraftingMonitor.getY(stack), ItemWirelessCraftingMonitor.getZ(stack));
         this.size = ItemWirelessCraftingMonitor.getSize(stack);
+        this.tabPage = ItemWirelessCraftingMonitor.getTabPage(stack);
+        this.tabSelected = ItemWirelessCraftingMonitor.getTabSelected(stack);
     }
 
     @Override
@@ -41,7 +47,7 @@ public class WirelessCraftingMonitor implements ICraftingMonitor {
     }
 
     @Override
-    public void onCancelled(EntityPlayerMP player, int id) {
+    public void onCancelled(EntityPlayerMP player, UUID id) {
         INetwork network = getNetwork();
 
         if (network != null) {
@@ -55,7 +61,7 @@ public class WirelessCraftingMonitor implements ICraftingMonitor {
     }
 
     @Override
-    public List<ICraftingTask> getTasks() {
+    public Collection<ICraftingTask> getTasks() {
         INetwork network = getNetwork();
 
         if (network != null) {
@@ -88,7 +94,7 @@ public class WirelessCraftingMonitor implements ICraftingMonitor {
 
         GuiBase.executeLater(GuiCraftingMonitor.class, GuiBase::initGui);
 
-        RS.INSTANCE.network.sendToServer(new MessageWirelessCraftingMonitorSize(size));
+        RS.INSTANCE.network.sendToServer(new MessageWirelessCraftingMonitorSettings(size, tabSelected, tabPage));
     }
 
     private INetwork getNetwork() {
@@ -118,6 +124,36 @@ public class WirelessCraftingMonitor implements ICraftingMonitor {
 
         if (network != null) {
             network.getNetworkItemHandler().onClose(player);
+        }
+    }
+
+    @Override
+    public Optional<UUID> getTabSelected() {
+        return tabSelected;
+    }
+
+    @Override
+    public int getTabPage() {
+        return tabPage;
+    }
+
+    @Override
+    public void onTabSelectionChanged(Optional<UUID> taskId) {
+        if (taskId.isPresent() && tabSelected.isPresent() && taskId.get().equals(tabSelected.get())) {
+            this.tabSelected = Optional.absent();
+        } else {
+            this.tabSelected = taskId;
+        }
+
+        RS.INSTANCE.network.sendToServer(new MessageWirelessCraftingMonitorSettings(size, tabSelected, tabPage));
+    }
+
+    @Override
+    public void onTabPageChanged(int page) {
+        if (page >= 0) {
+            this.tabPage = page;
+
+            RS.INSTANCE.network.sendToServer(new MessageWirelessCraftingMonitorSettings(size, tabSelected, tabPage));
         }
     }
 }